Split up tanjun.context.py into directory
