+++ tanjun/context/slash.py
@@ -29,10 +29,10 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-"""Standard command execution context implementations."""
+"""Standard slash command execution context implementations."""
 from __future__ import annotations
 
-__all__: list[str] = ["MessageContext", "ResponseTypeT", "SlashContext", "SlashOption"]
+__all__: list[str] = ["SlashContext", "SlashOption"]
 
 import asyncio
 import datetime
@@ -40,498 +40,25 @@ import logging
 import typing
 
 import hikari
-from hikari import snowflakes
 
-from . import abc as tanjun_abc
-from . import injecting
+from .. import abc as tanjun_abc
+from . import base
 
 if typing.TYPE_CHECKING:
     from collections import abc as collections
 
-    from hikari import traits as hikari_traits
+    from .. import injecting
 
-    _BaseContextT = typing.TypeVar("_BaseContextT", bound="BaseContext")
-    _MessageContextT = typing.TypeVar("_MessageContextT", bound="MessageContext")
     _SlashContextT = typing.TypeVar("_SlashContextT", bound="SlashContext")
     _T = typing.TypeVar("_T")
 
-ResponseTypeT = typing.Union[hikari.api.InteractionMessageBuilder, hikari.api.InteractionDeferredBuilder]
-"""Union of the response types which are valid for application command interactions."""
+_ResponseTypeT = typing.Union[hikari.api.InteractionMessageBuilder, hikari.api.InteractionDeferredBuilder]
 _INTERACTION_LIFETIME: typing.Final[datetime.timedelta] = datetime.timedelta(minutes=15)
 _LOGGER = logging.getLogger("hikari.tanjun.context")
 
 
 def _delete_after_to_float(delete_after: typing.Union[datetime.timedelta, float, int]) -> float:
     return delete_after.total_seconds() if isinstance(delete_after, datetime.timedelta) else float(delete_after)
-
-
-class BaseContext(injecting.BasicInjectionContext, tanjun_abc.Context):
-    """Base class for all standard context implementations."""
-
-    __slots__ = ("_client", "_component", "_final")
-
-    def __init__(
-        self,
-        client: tanjun_abc.Client,
-        injection_client: injecting.InjectorClient,
-        *,
-        component: typing.Optional[tanjun_abc.Component] = None,
-    ) -> None:
-        # injecting.BasicInjectionContext.__init__
-        super().__init__(injection_client)
-        self._client = client
-        self._component = component
-        self._final = False
-        (
-            self._set_type_special_case(tanjun_abc.Context, self)
-            ._set_type_special_case(BaseContext, self)
-            ._set_type_special_case(type(self), self)
-        )
-
-    @property
-    def cache(self) -> typing.Optional[hikari.api.Cache]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client.cache
-
-    @property
-    def client(self) -> tanjun_abc.Client:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client
-
-    @property
-    def component(self) -> typing.Optional[tanjun_abc.Component]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._component
-
-    @property
-    def events(self) -> typing.Optional[hikari.api.EventManager]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client.events
-
-    @property
-    def server(self) -> typing.Optional[hikari.api.InteractionServer]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client.server
-
-    @property
-    def rest(self) -> hikari.api.RESTClient:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client.rest
-
-    @property
-    def shards(self) -> typing.Optional[hikari_traits.ShardAware]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client.shards
-
-    @property
-    def voice(self) -> typing.Optional[hikari.api.VoiceComponent]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._client.voice
-
-    def _assert_not_final(self) -> None:
-        if self._final:
-            raise TypeError("Cannot modify a finalised context")
-
-    def finalise(self: _BaseContextT) -> _BaseContextT:
-        """Finalise the context, dis-allowing any further modifications.
-
-        Returns
-        -------
-        Self
-            The context itself to enable chained calls.
-        """
-        self._final = True
-        return self
-
-    def set_component(self: _BaseContextT, component: typing.Optional[tanjun_abc.Component], /) -> _BaseContextT:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        self._assert_not_final()
-        if component:
-            self._set_type_special_case(tanjun_abc.Component, component)._set_type_special_case(
-                type(component), component
-            )
-
-        elif component_case := self._special_case_types.get(tanjun_abc.Component):
-            self._remove_type_special_case(tanjun_abc.Component)
-            self._remove_type_special_case(type(component_case))
-
-        self._component = component
-        return self
-
-    def get_channel(self) -> typing.Optional[hikari.TextableGuildChannel]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self._client.cache:
-            channel = self._client.cache.get_guild_channel(self.channel_id)
-            assert channel is None or isinstance(channel, hikari.TextableGuildChannel)
-            return channel
-
-        return None
-
-    def get_guild(self) -> typing.Optional[hikari.Guild]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self.guild_id is not None and self._client.cache:
-            return self._client.cache.get_guild(self.guild_id)
-
-        return None
-
-    async def fetch_channel(self) -> hikari.TextableChannel:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        channel = await self._client.rest.fetch_channel(self.channel_id)
-        assert isinstance(channel, hikari.TextableChannel)
-        return channel
-
-    async def fetch_guild(self) -> typing.Optional[hikari.Guild]:  # TODO: or raise?
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self.guild_id is not None:
-            return await self._client.rest.fetch_guild(self.guild_id)
-
-        return None
-
-
-class MessageContext(BaseContext, tanjun_abc.MessageContext):
-    """Standard implementation of a command context as used within Tanjun."""
-
-    __slots__ = (
-        "_command",
-        "_content",
-        "_initial_response_id",
-        "_last_response_id",
-        "_response_lock",
-        "_message",
-        "_triggering_name",
-        "_triggering_prefix",
-    )
-
-    def __init__(
-        self,
-        client: tanjun_abc.Client,
-        injection_client: injecting.InjectorClient,
-        content: str,
-        message: hikari.Message,
-        *,
-        command: typing.Optional[tanjun_abc.MessageCommand[typing.Any]] = None,
-        component: typing.Optional[tanjun_abc.Component] = None,
-        triggering_name: str = "",
-        triggering_prefix: str = "",
-    ) -> None:
-        if message.content is None:
-            raise ValueError("Cannot spawn context with a content-less message.")
-
-        super().__init__(client, injection_client, component=component)
-        self._command = command
-        self._content = content
-        self._initial_response_id: typing.Optional[hikari.Snowflake] = None
-        self._last_response_id: typing.Optional[hikari.Snowflake] = None
-        self._response_lock = asyncio.Lock()
-        self._message = message
-        self._triggering_name = triggering_name
-        self._triggering_prefix = triggering_prefix
-        self._set_type_special_case(tanjun_abc.MessageContext, self)._set_type_special_case(MessageContext, self)
-
-    def __repr__(self) -> str:
-        return f"MessageContext <{self._message!r}, {self._command!r}>"
-
-    @property
-    def author(self) -> hikari.User:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._message.author
-
-    @property
-    def channel_id(self) -> hikari.Snowflake:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._message.channel_id
-
-    @property
-    def command(self) -> typing.Optional[tanjun_abc.MessageCommand[typing.Any]]:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        return self._command
-
-    @property
-    def content(self) -> str:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        return self._content
-
-    @property
-    def created_at(self) -> datetime.datetime:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._message.created_at
-
-    @property
-    def guild_id(self) -> typing.Optional[hikari.Snowflake]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._message.guild_id
-
-    @property
-    def has_responded(self) -> bool:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._initial_response_id is not None
-
-    @property
-    def is_human(self) -> bool:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return not self._message.author.is_bot and self._message.webhook_id is None
-
-    @property
-    def member(self) -> typing.Optional[hikari.Member]:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._message.member
-
-    @property
-    def message(self) -> hikari.Message:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        return self._message
-
-    @property
-    def triggering_name(self) -> str:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        return self._triggering_name
-
-    @property
-    def triggering_prefix(self) -> str:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        return self._triggering_prefix
-
-    @property
-    def shard(self) -> typing.Optional[hikari.api.GatewayShard]:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        if not self._client.shards:
-            return None
-
-        if self._message.guild_id is not None:
-            shard_id = snowflakes.calculate_shard_id(self._client.shards, self._message.guild_id)
-
-        else:
-            shard_id = 0
-
-        return self._client.shards.shards[shard_id]
-
-    def set_command(
-        self: _MessageContextT, command: typing.Optional[tanjun_abc.MessageCommand[typing.Any]], /
-    ) -> _MessageContextT:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        self._assert_not_final()
-        self._command = command
-        if command:
-            (
-                self._set_type_special_case(tanjun_abc.ExecutableCommand, command)
-                ._set_type_special_case(tanjun_abc.MessageCommand, command)
-                ._set_type_special_case(type(command), command)
-            )
-
-        elif command_case := self._special_case_types.get(tanjun_abc.ExecutableCommand):
-            self._remove_type_special_case(tanjun_abc.ExecutableCommand)
-            self._remove_type_special_case(tanjun_abc.MessageCommand)  # TODO: command group?
-            self._remove_type_special_case(type(command_case))
-
-        return self
-
-    def set_content(self: _MessageContextT, content: str, /) -> _MessageContextT:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        self._assert_not_final()
-        self._content = content
-        return self
-
-    def set_triggering_name(self: _MessageContextT, name: str, /) -> _MessageContextT:
-        # <<inherited docstring from tanjun.abc.MessageContext>>.
-        self._assert_not_final()
-        self._triggering_name = name
-        return self
-
-    def set_triggering_prefix(self: _MessageContextT, triggering_prefix: str, /) -> _MessageContextT:
-        """Set the triggering prefix for this context.
-
-        Parameters
-        ----------
-        triggering_prefix : str
-            The triggering prefix to set.
-
-        Returns
-        -------
-        Self
-            This context to allow for chaining.
-        """
-        self._assert_not_final()
-        self._triggering_prefix = triggering_prefix
-        return self
-
-    async def delete_initial_response(self) -> None:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self._initial_response_id is None:
-            raise LookupError("Context has no initial response")
-
-        await self._client.rest.delete_message(self._message.channel_id, self._initial_response_id)
-
-    async def delete_last_response(self) -> None:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self._last_response_id is None:
-            raise LookupError("Context has no previous responses")
-
-        await self._client.rest.delete_message(self._message.channel_id, self._last_response_id)
-
-    async def edit_initial_response(
-        self,
-        content: hikari.UndefinedOr[typing.Any] = hikari.UNDEFINED,
-        *,
-        delete_after: typing.Union[datetime.timedelta, float, int, None] = None,
-        attachment: hikari.UndefinedOr[hikari.Resourceish] = hikari.UNDEFINED,
-        attachments: hikari.UndefinedOr[collections.Sequence[hikari.Resourceish]] = hikari.UNDEFINED,
-        component: hikari.UndefinedNoneOr[hikari.api.ComponentBuilder] = hikari.UNDEFINED,
-        components: hikari.UndefinedNoneOr[collections.Sequence[hikari.api.ComponentBuilder]] = hikari.UNDEFINED,
-        embed: hikari.UndefinedNoneOr[hikari.Embed] = hikari.UNDEFINED,
-        embeds: hikari.UndefinedNoneOr[collections.Sequence[hikari.Embed]] = hikari.UNDEFINED,
-        replace_attachments: bool = False,
-        mentions_everyone: hikari.UndefinedOr[bool] = hikari.UNDEFINED,
-        user_mentions: typing.Union[
-            hikari.SnowflakeishSequence[hikari.PartialUser], bool, hikari.UndefinedType
-        ] = hikari.UNDEFINED,
-        role_mentions: typing.Union[
-            hikari.SnowflakeishSequence[hikari.PartialRole], bool, hikari.UndefinedType
-        ] = hikari.UNDEFINED,
-    ) -> hikari.Message:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        delete_after = _delete_after_to_float(delete_after) if delete_after is not None else None
-        if self._initial_response_id is None:
-            raise LookupError("Context has no initial response")
-
-        message = await self.rest.edit_message(
-            self._message.channel_id,
-            self._initial_response_id,
-            content=content,
-            attachment=attachment,
-            attachments=attachments,
-            component=component,
-            components=components,
-            embed=embed,
-            embeds=embeds,
-            replace_attachments=replace_attachments,
-            mentions_everyone=mentions_everyone,
-            user_mentions=user_mentions,
-            role_mentions=role_mentions,
-        )
-        if delete_after is not None:
-            asyncio.create_task(self._delete_after(delete_after, message))
-
-        return message
-
-    async def edit_last_response(
-        self,
-        content: hikari.UndefinedOr[typing.Any] = hikari.UNDEFINED,
-        *,
-        delete_after: typing.Union[datetime.timedelta, float, int, None] = None,
-        attachment: hikari.UndefinedOr[hikari.Resourceish] = hikari.UNDEFINED,
-        attachments: hikari.UndefinedOr[collections.Sequence[hikari.Resourceish]] = hikari.UNDEFINED,
-        component: hikari.UndefinedNoneOr[hikari.api.ComponentBuilder] = hikari.UNDEFINED,
-        components: hikari.UndefinedNoneOr[collections.Sequence[hikari.api.ComponentBuilder]] = hikari.UNDEFINED,
-        embed: hikari.UndefinedNoneOr[hikari.Embed] = hikari.UNDEFINED,
-        embeds: hikari.UndefinedNoneOr[collections.Sequence[hikari.Embed]] = hikari.UNDEFINED,
-        replace_attachments: bool = False,
-        mentions_everyone: hikari.UndefinedOr[bool] = hikari.UNDEFINED,
-        user_mentions: typing.Union[
-            hikari.SnowflakeishSequence[hikari.PartialUser], bool, hikari.UndefinedType
-        ] = hikari.UNDEFINED,
-        role_mentions: typing.Union[
-            hikari.SnowflakeishSequence[hikari.PartialRole], bool, hikari.UndefinedType
-        ] = hikari.UNDEFINED,
-    ) -> hikari.Message:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        delete_after = _delete_after_to_float(delete_after) if delete_after is not None else None
-        if self._last_response_id is None:
-            raise LookupError("Context has no previous tracked response")
-
-        message = await self.rest.edit_message(
-            self._message.channel_id,
-            self._last_response_id,
-            content=content,
-            attachment=attachment,
-            attachments=attachments,
-            component=component,
-            components=components,
-            embed=embed,
-            embeds=embeds,
-            replace_attachments=replace_attachments,
-            mentions_everyone=mentions_everyone,
-            user_mentions=user_mentions,
-            role_mentions=role_mentions,
-        )
-
-        if delete_after is not None:
-            asyncio.create_task(self._delete_after(delete_after, message))
-
-        return message
-
-    async def fetch_initial_response(self) -> hikari.Message:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self._initial_response_id is not None:
-            return await self.client.rest.fetch_message(self._message.channel_id, self._initial_response_id)
-
-        raise LookupError("No initial response found for this context")
-
-    async def fetch_last_response(self) -> hikari.Message:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        if self._last_response_id is not None:
-            return await self.client.rest.fetch_message(self._message.channel_id, self._last_response_id)
-
-        raise LookupError("No responses found for this context")
-
-    @staticmethod
-    async def _delete_after(delete_after: float, message: hikari.Message) -> None:
-        await asyncio.sleep(delete_after)
-        try:
-            await message.delete()
-        except hikari.NotFoundError as exc:
-            _LOGGER.debug("Failed to delete response message after %.2f seconds", delete_after, exc_info=exc)
-
-    async def respond(
-        self,
-        content: hikari.UndefinedOr[typing.Any] = hikari.UNDEFINED,
-        *,
-        ensure_result: bool = True,
-        delete_after: typing.Union[datetime.timedelta, float, int, None] = None,
-        attachment: hikari.UndefinedOr[hikari.Resourceish] = hikari.UNDEFINED,
-        attachments: hikari.UndefinedOr[collections.Sequence[hikari.Resourceish]] = hikari.UNDEFINED,
-        component: hikari.UndefinedOr[hikari.api.ComponentBuilder] = hikari.UNDEFINED,
-        components: hikari.UndefinedOr[collections.Sequence[hikari.api.ComponentBuilder]] = hikari.UNDEFINED,
-        embed: hikari.UndefinedOr[hikari.Embed] = hikari.UNDEFINED,
-        embeds: hikari.UndefinedOr[collections.Sequence[hikari.Embed]] = hikari.UNDEFINED,
-        tts: hikari.UndefinedOr[bool] = hikari.UNDEFINED,
-        nonce: hikari.UndefinedOr[str] = hikari.UNDEFINED,
-        reply: typing.Union[bool, hikari.SnowflakeishOr[hikari.PartialMessage], hikari.UndefinedType] = False,
-        mentions_everyone: hikari.UndefinedOr[bool] = hikari.UNDEFINED,
-        mentions_reply: hikari.UndefinedOr[bool] = hikari.UNDEFINED,
-        user_mentions: typing.Union[
-            hikari.SnowflakeishSequence[hikari.PartialUser], bool, hikari.UndefinedType
-        ] = hikari.UNDEFINED,
-        role_mentions: typing.Union[
-            hikari.SnowflakeishSequence[hikari.PartialRole], bool, hikari.UndefinedType
-        ] = hikari.UNDEFINED,
-    ) -> hikari.Message:
-        # <<inherited docstring from tanjun.abc.Context>>.
-        delete_after = _delete_after_to_float(delete_after) if delete_after is not None else None
-        async with self._response_lock:
-            message = await self._message.respond(
-                content=content,
-                attachment=attachment,
-                attachments=attachments,
-                component=component,
-                components=components,
-                embed=embed,
-                embeds=embeds,
-                tts=tts,
-                nonce=nonce,
-                reply=reply,
-                mentions_everyone=mentions_everyone,
-                mentions_reply=mentions_reply,
-                user_mentions=user_mentions,
-                role_mentions=role_mentions,
-            )
-            self._last_response_id = message.id
-            if self._initial_response_id is None:
-                self._initial_response_id = message.id
-
-            if delete_after is not None:
-                asyncio.create_task(self._delete_after(delete_after, message))
-
-            return message
 
 
 _SnowflakeOptions = {
@@ -734,7 +261,7 @@ _COMMAND_OPTION_TYPES: typing.Final[frozenset[hikari.OptionType]] = frozenset(
 )
 
 
-class SlashContext(BaseContext, tanjun_abc.SlashContext):
+class SlashContext(base.BaseContext, tanjun_abc.SlashContext):
     __slots__ = (
         "_command",
         "_defaults_to_ephemeral",
@@ -771,7 +298,7 @@ class SlashContext(BaseContext, tanjun_abc.SlashContext):
         self._last_response_id: typing.Optional[hikari.Snowflake] = None
         self._marked_not_found = False
         self._on_not_found = on_not_found
-        self._response_future: typing.Optional[asyncio.Future[ResponseTypeT]] = None
+        self._response_future: typing.Optional[asyncio.Future[_ResponseTypeT]] = None
         self._response_lock = asyncio.Lock()
         self._set_type_special_case(tanjun_abc.SlashContext, self)._set_type_special_case(SlashContext, self)
 
@@ -878,7 +405,7 @@ class SlashContext(BaseContext, tanjun_abc.SlashContext):
 
         return flags or hikari.MessageFlag.NONE
 
-    def get_response_future(self) -> asyncio.Future[ResponseTypeT]:
+    def get_response_future(self) -> asyncio.Future[_ResponseTypeT]:
         """Get the future which will be used to set the initial response.
 
         .. note::
@@ -887,7 +414,7 @@ class SlashContext(BaseContext, tanjun_abc.SlashContext):
 
         Returns
         -------
-        asyncio.Future[ResponseTypeT]
+        asyncio.Future[_ResponseTypeT]
             The future which will be used to set the initial response.
         """
         if not self._response_future:
